// src/api/generate.ts
export class CreateOrderModel implements ICreateOrderModel {
    customer?: string | undefined;
    description?: string | undefined;
    costs!: number;

    constructor(data?: ICreateOrderModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.customer = _data["customer"];
            this.description = _data["description"];
            this.costs = _data["costs"];
        }
    }

    static fromJS(data: any): CreateOrderModel {
        data = typeof data === 'object' ? data : {};
        let result = new CreateOrderModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["customer"] = this.customer;
        data["description"] = this.description;
        data["costs"] = this.costs;
        return data; 
    }
}

export interface ICreateOrderModel {
    customer?: string | undefined;
    description?: string | undefined;
    costs: number;
}